gpgme: refactoring - move keylist to separate module
